+++ esmerald_admin/backends/base.py
@@ -2,17 +2,16 @@ from datetime import datetime, timedelta
 from typing import Any, Optional
 
 from esmerald import Request
-from esmerald.exceptions import AuthenticationError, NotAuthorized
+from esmerald.exceptions import AuthenticationError
 from esmerald.security.jwt.token import Token
 from jose import JWSError, JWTError
-from saffier.exceptions import DoesNotFound
 from sqladmin.authentication import AuthenticationBackend
 from starlette.responses import RedirectResponse
 
 DEFAULT_HEADER = "Bearer"
 
 
-class BaseAuthentication(AuthenticationBackend):
+class BackendBaseAuthentication(AuthenticationBackend):
     """
     Uses the AuthenticationProtocol from esmerald_admin assuming it is using the
     Esmerald contrib user and login into the admin.
@@ -23,7 +22,7 @@ class BaseAuthentication(AuthenticationBackend):
         self.auth_model = auth_model
         self.config = config
 
-    def generate_user_token(self, user: Any, time=None):
+    def generate_user_token(self, user: Any, time: Any = None) -> str:
         """
         Generates the JWT token for the authenticated user.
         """
@@ -35,14 +34,14 @@ class BaseAuthentication(AuthenticationBackend):
         token = Token(sub=user.id, exp=later)
         return token.encode(key=self.config.signing_key, algorithm=self.config.algorithm)
 
-    def is_user_able_to_authenticate(self, user):
+    def is_user_able_to_authenticate(self, user: Any) -> bool:
         """
         Reject users with is_active=False. Custom user models that don't have
         that attribute are allowed.
         """
         return getattr(user, "is_active", True)
 
-    def is_user_staff_and_superuser(self, user):
+    def is_user_staff_and_superuser(self, user: Any) -> bool:
         """Checks if a user is staff and superuser to acess the admin"""
         return bool(user.is_staff and user.is_superuser)
 
@@ -54,24 +53,6 @@ class BaseAuthentication(AuthenticationBackend):
         """Logout from the admin"""
         await self.clear_session(request)
         return True
-
-    async def retrieve_user(self, token_sub: Any) -> Any:
-        """
-        Retrieves a user from the database using the given token id.
-        """
-        try:
-            sub = int(token_sub)
-            token_sub = sub
-        except (TypeError, ValueError):
-            ...
-
-        user_field = {self.config.user_id_field: token_sub}
-        try:
-            return await self.auth_model.query.get(**user_field)
-        except DoesNotFound:
-            raise NotAuthorized() from None
-        except Exception as e:
-            raise AuthenticationError(detail=str(e)) from e
 
     async def authenticate(self, request: Request) -> Optional[RedirectResponse]:  # type: ignore
         """Authenticates the user and adds to the scope of the application"""