+++ docs/release-notes.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## 0.3.0
+
+### Changed
+
+- Updated requirement versions.
+- **Breaking change** - The installation does not default to install Saffier anymore. For that purpose
+you will need to specify now `pip install esmerald-admin[saffier]` or `pip install esmerald-admin[edgy]`.
+
+### Added
+
+- Support for [Edgy](https://edgy.tarsild.io).
+
 ## 0.2.0
 
 ### Changed